agrego nuevas funciones en los usuarios

+++ js/admin.js
@@ -166,6 +166,7 @@ const cargarTablaUser = () => {
         <th>
         <div class="d-flex gap-2">
         <i class="fa-solid fa-arrow-up puntero" onclick="hacerAdmin(${users.id})" style="color: #04ff00;"></i>
+        <i class="fa-solid fa-arrow-down puntero" onclick="hacerUsuario(${users.id} )" style="color: #d01616;"></i>
         <i class="fa-solid fa-ban puntero" onclick="eliminarUser(${users.id} )" style="color: #ff0000;"></i>
         </div>
         </th>
@@ -181,10 +182,38 @@ const hacerAdmin = (id) =>{
     let indice = usuarios.findIndex((item)=>{
         return item.id == id;
       });
-      usuarios[indice].admin = !usuarios[indice].admin;
+      if (usuarios[indice].admin) {
+        alert("El usuario ya es administrador")
+      }else{
+        let validar = confirm("Seguro que desea ascender a admin?")
+
+        if (validar) {
+          usuarios[indice].admin = true;
+      
+          localStorage.setItem("usuarios",JSON.stringify(usuarios)); 
+        }
+      }
     
-      localStorage.setItem("usuarios",JSON.stringify(usuarios));
-    
+      cargarTablaUser();
+}
+
+const hacerUsuario = (id) =>{
+    let indice = usuarios.findIndex((item)=>{
+        return item.id == id;
+      });
+
+      if (!usuarios[indice].admin) {
+        alert("El usuario ya es user")
+      }else{
+        let validar = confirm("Seguro que desea descender a user?")
+        if (validar) {
+          usuarios[indice].admin = false;
+      
+          localStorage.setItem("usuarios",JSON.stringify(usuarios));
+        }
+      }
+
+
       cargarTablaUser();
 }
